Patch from Steve Cohen:
This is a series of mostly cosmetic changes in response to the CheckStyle
suggestions and a couple of howling pieces of debugging junk that should not
have made it through - System.out.println() and the like.  A little more
significantly, two instances where exceptions were eaten were fixed -
in FTPFileList and MatchApparatus.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/net/trunk@139300 13f79535-47bb-0310-9956-ffa450edef68