Submitted by: Tapan Karecha <tapan@india.hp.com> Reviewed by: dfs
Applied patch for restarting FTP file transfers.  The offset was not
being sent immediately before the data transfer command on account.
The bug was apparently introduced in NetComponents when it was decided
to always send a PORT command before each data transfer to avoid
socket reuse problems on Windows.  Tapan Karecha detected the error
and supplied this fix.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/net/trunk@139327 13f79535-47bb-0310-9956-ffa450edef68