Fixed the following problems reported by Stephen Colebourne: > [net] currently contains two packages > o.a.commons.io > o.a.commons.util > These packages need to be renamed to be > o.a.commons.net.io > o.a.commons.net.util > > Also, some of the licences in the source files refer to the turbine licence, > not the commons licence. This should be changed.
Also commented out the dependency of the jar build rule on the test build
rule because the test rule doesn't work for me and if it doesn't work for
me, it won't work for the average person who checks out the source.  Will
resolve later (probably punt to Jeff since I've yet to learn Maven).


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/net/trunk@139336 13f79535-47bb-0310-9956-ffa450edef68