Applied patch from Paul Spencer <paul@mikon.com>: >Please apply the attached patch.  It corrects the following problems: > >1) toString() failed in JDK 1.3 because the method >StringBuffer.append(StringBuffer) does not exist.  The method was added >in JDK 1.4. > >2) Empty subjects are in the message.  Now the subject will only be >includes when it is non-null.
What appears to have happened is that the code relied on an implicit
toString() call by the compiler, but after J2SE 1.4 added
StringBuffer.append(StringBuffer) and the code was compiled using
J2SE 1.4 javac, a call to StringBuffer.append(StringBuffer) was
resolved instead of StringBuffer.append(String).  Paul's patch
fixes this problem.  Other parts of the code that depended on an
implicit StringBuffer.toString() may be affected as well and should
be checked.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/net/trunk@139386 13f79535-47bb-0310-9956-ffa450edef68