What I tried to do was replace the references to straight up files with {@link classname} or {@link classname#member} and references to in file anchors with {@link #member}.  But there were quite a number of classname.html#_top_ references and I replaced those with {@link classname}. So I ended up running the following:  find -name "*.java" | xargs perl -pi -e  's/<a\s+href="([^"]+)\.html(?:#_top_)?">[^<]+<\/a>/{\@link $1}/g;  s/<a\s+href="([^"]+)\.html(#[^"]+)">([^<]+)<\/a>/{\@link $1$2 $3}/g;  s/<a\s+href="(#[^"]+)">([^<]+)<\/a>/{\@link $1 $2}/g'
I doublechecked that everything still compiled fine and that 'maven
javadoc'
didn't choke.  javadoc spouted warnings because there were typos in
the
some of the links (e.g., sendComand instead of sendCommand).  So I fixed
those with:
 find -name "*.java" | xargs perl -pi -e 's/Comand/Command/g'
 find -name "*.java" | xargs perl -pi -e 's/recieve/receive/g'
There were some mistaken references to sendMessage which should have
been
sendMessageData in SMTPClient, so I fixed those similarly.

Lastly, FTPClient had a reference to Package-packagename.html which were
translated into @link Package-packagename, so I trimmed off the Package-,
leaving {@link org.apache.commons.net.io}, which javadoc complained
about with:
  warning - Tag @link: reference not found: org.apache.commons.net.io
This is supposed to work, so I left it alone because the only damage
done
to the docs is that the package name is printed without a hyperlink.

There are still a bunch of .html links that span across two javadoc
lines and I'll fix those next, but wanted to report on the initial fixes and
check them in.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/net/trunk@139515 13f79535-47bb-0310-9956-ffa450edef68